HSEARCH-1375 Move CountingFailureHandler (now StubFailureHandler) to test utils
